T2/cache_system/requirements.txt
redis
psycopg2-binary
kafka-python==2.0.2